blacklight marc document override, remove spelling override
