+++ app/assets/stylesheets/blacklight.scss
@@ -5,5 +5,5 @@ $table-bg-accent: #f5f4f1;
 @import 'bootstrap';
 
 @import 'blacklight/blacklight';
-// @import 'blacklight_folders/blacklight_folders';
 @import 'blacklight/blacklight_base';
+@import 'blacklight_marc';
